exercises/proyecciones-cubo3d/requirements.txt
glfw==2.7.0
PyOpenGL==3.1.7
pyopengltk==0.0.4
